Code compiles.  Analog readings function.  Features have been added to the console.  Some cleanup has occurred.

+++ digital_hygrometer_esp32_c3/i2c.h
@@ -152,7 +152,7 @@ class I2C {
          * @return nothing 
          * 
          */ 
-        void choose_sensor(int sensor_number);
+        void choose_sensor(uint8_t sensor_number);
 
         //TODO: need to comment
         void configure_sensor( void );
